+++ src/components/customForm/CustomFormEdit.tsx
@@ -1,23 +1,36 @@
-import { useAlert, useDataEngine } from '@dhis2/app-runtime'
+import { useAlert } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
 import { Button, ButtonStrip, NoticeBox } from '@dhis2/ui'
 import { IconInfo16 } from '@dhis2/ui-icons'
-import React, { useCallback, useEffect, useRef, useState } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import { useField } from 'react-final-form'
-import { useParams } from 'react-router-dom'
 import {
     FormFooterWrapper,
+    SectionedFormErrorNotice,
     SectionedFormSection,
     SectionedFormSections,
     StandardFormSectionDescription,
     StandardFormSectionTitle,
 } from '..'
-import { generateDhis2Id, useSectionHandle } from '../../lib'
+import { generateDhis2Id } from '../../lib'
 import styles from './CustomFormContents.module.css'
-import { CustomFormElementsSelectorJunction } from './CustomFormElementsSelector'
+import {
+    LoadingCustomFormElementsSelector,
+    ElementTypes,
+} from './CustomFormElementsSelector'
 
 export type CustomFormProps = {
     closeCustomFormEdit?: () => void
+    loading: boolean
+    refetch: () => void
+    elementTypes: ElementTypes
+    updateCustomForm: (
+        data: CustomFormDataPayload,
+        onSuccess: (data: CustomFormDataPayload) => void,
+        onError: (e: Error) => void,
+        existingFormId: string | undefined
+    ) => Promise<unknown>
+    customFormTarget: string
 }
 
 const SubsectionSpacer = ({ children }: { children: React.ReactNode }) => (
@@ -36,10 +49,12 @@ const getElementText = ({
     disabled?: boolean
 }): string => {
     if (type === 'dataElement') {
-        const [de, coc] = id.split('.')
+        const [de, parent] = id.split('.')
+        const formattedId =
+            parent === undefined ? `${de}-val` : `${de}-${parent}-val`
         return `<p><input ${
             disabled ? 'disabled="disabled"' : ''
-        } id="${de}-${coc}-val" name="entryfield" title="${name}" value="[ ${name} ]" /></p>`
+        } id="${formattedId}" name="entryfield" title="${name}" value="[ ${name} ]" /></p>`
     }
     if (type === 'total') {
         return `<p><input dataelementid="${id}" ${
@@ -63,146 +78,26 @@ const getElementText = ({
     return ''
 }
 
-type CustomFormDataPayload = {
+export type CustomFormDataPayload = {
     id: string
     htmlCode: string
     name?: string
 }
-// when there is a data set id, we can post to the form endpoint
-const useUpdateForm = ({
-    onSuccess,
-    onError,
-    parentSchema,
-    operationType,
-}: {
-    onSuccess: (data: CustomFormDataPayload) => void
-    onError: (e: Error) => void
-    parentSchema: 'dataSet' | 'program'
-    operationType: 'create' | 'update'
-}) => {
-    const id = useParams().id
-    const dataEngine = useDataEngine()
 
-    const updateDataSet = useCallback(
-        async (data: CustomFormDataPayload) => {
-            try {
-                const response = await dataEngine.mutate(
-                    {
-                        resource: `dataSets/${id}/form`,
-                        type: 'create',
-                        data: data,
-                    },
-                    {
-                        onComplete: () => {
-                            // the response from this post is empty, so we use the data we passed if it was successful
-                            onSuccess(data)
-                        },
-                        onError,
-                    }
-                )
-                return { data: response }
-            } catch (error) {
-                console.error(error)
-            }
-        },
-        [dataEngine, id, onSuccess, onError]
-    )
-
-    const updateProgram = useCallback(
-        async (data: CustomFormDataPayload) => {
-            try {
-                const response = await dataEngine.mutate(
-                    {
-                        resource: `dataEntryForms`,
-                        id: data.id,
-                        type: 'json-patch',
-                        data: [
-                            {
-                                op: 'replace',
-                                path: '/htmlCode',
-                                value: data.htmlCode,
-                            },
-                        ],
-                    },
-                    {
-                        onComplete: () => {
-                            // the response from this post is empty, so we use the data we passed if it was successful
-                            onSuccess(data)
-                        },
-                        onError,
-                    }
-                )
-                return { data: response }
-            } catch (error) {
-                console.error(error)
-            }
-        },
-        [dataEngine, onSuccess, onError]
-    )
-
-    const createProgram = useCallback(
-        async (data: CustomFormDataPayload) => {
-            try {
-                const response = await dataEngine.mutate(
-                    {
-                        resource: `dataEntryForms`,
-                        type: 'create',
-                        data: data,
-                    },
-                    {
-                        onError,
-                    }
-                )
-                await dataEngine.mutate(
-                    {
-                        resource: `programs`,
-                        id: id as string,
-                        type: 'json-patch',
-                        data: [
-                            {
-                                op: 'replace',
-                                path: '/dataEntryForm',
-                                value: { id: data.id },
-                            },
-                        ],
-                    },
-                    {
-                        onComplete: () => {
-                            // use the data we passed if form was saved and associated to program
-                            onSuccess(data)
-                        },
-                    }
-                )
-                return { data: response }
-            } catch (error) {
-                console.error(error)
-            }
-        },
-        [dataEngine, id, onSuccess, onError]
-    )
-    if (parentSchema === 'dataSet') {
-        return updateDataSet
-    }
-    if (operationType === 'create') {
-        return createProgram
-    }
-
-    return updateProgram
-}
-
-export const CustomFormEdit = ({ closeCustomFormEdit }: CustomFormProps) => {
+export const CustomFormEdit = ({
+    closeCustomFormEdit,
+    loading,
+    refetch,
+    elementTypes,
+    updateCustomForm,
+    customFormTarget,
+}: CustomFormProps) => {
     const textAreaRef = useRef<HTMLTextAreaElement>(null)
     const [previewMode, setPreviewMode] = useState<boolean>(false)
     const [customFormSaving, setCustomFormSaving] = useState<boolean>(false)
     const [customFormError, setCustomFormError] = useState<string>('')
     const { input: formInput } = useField('dataEntryForm')
     const { input: nameInput } = useField('name')
-    const section = useSectionHandle()
-    const customFormTarget =
-        section?.name === 'dataSet'
-            ? i18n.t('data set')
-            : i18n.t('program enrollment')
-    const operationType = formInput?.value?.id ? 'update' : 'create'
 
     const insertElement = ({
         id,
@@ -230,6 +125,10 @@ export const CustomFormEdit = ({ closeCustomFormEdit }: CustomFormProps) => {
     }
 
     useEffect(() => {
+        refetch()
+    }, [refetch])
+
+    useEffect(() => {
         if (!textAreaRef.current) {
             return
         }
@@ -240,27 +139,24 @@ export const CustomFormEdit = ({ closeCustomFormEdit }: CustomFormProps) => {
         success: true,
     })
 
-    const updateCustomForm = useUpdateForm({
-        onSuccess: (data) => {
-            showSuccess({ success: true })
-            // update form state
-            formInput.onChange({
-                ...formInput.value,
-                ...data,
-            })
+    const onSuccess = (data: CustomFormDataPayload) => {
+        showSuccess({ success: true })
+        // update form state
+        formInput.onChange({
+            ...formInput.value,
+            ...data,
+        })
 
-            // close module
-            if (closeCustomFormEdit) {
-                closeCustomFormEdit()
-            }
-        },
-        onError: (e) => {
-            setCustomFormSaving(false)
-            setCustomFormError(e.message)
-        },
-        parentSchema: section?.name === 'dataSet' ? 'dataSet' : 'program',
-        operationType,
-    })
+        // close module
+        if (closeCustomFormEdit) {
+            closeCustomFormEdit()
+        }
+    }
+
+    const onError = (e: Error) => {
+        setCustomFormSaving(false)
+        setCustomFormError(e.message)
+    }
 
     return (
         <div className={styles.sectionsWrapper}>
@@ -329,14 +225,17 @@ export const CustomFormEdit = ({ closeCustomFormEdit }: CustomFormProps) => {
                                 ></textarea>
                             </div>
                             <div className={styles.customFormElementsContainer}>
-                                <CustomFormElementsSelectorJunction
+                                <LoadingCustomFormElementsSelector
                                     insertElement={insertElement}
                                     previewMode={previewMode}
+                                    loading={loading}
+                                    elementTypes={elementTypes}
                                 />
                             </div>
                         </div>
                     </SectionedFormSection>
                 </SectionedFormSections>
+                <SectionedFormErrorNotice />
                 {customFormError?.length > 0 && (
                     <div className={styles.errorNoticeWrapper}>
                         <NoticeBox
@@ -357,6 +256,8 @@ export const CustomFormEdit = ({ closeCustomFormEdit }: CustomFormProps) => {
                             primary
                             small
                             type="button"
+                            dataTest="form-submit-button"
+                            disabled={customFormSaving}
                             onClick={() => {
                                 const formId =
                                     formInput?.value?.id ?? generateDhis2Id()
@@ -364,21 +265,24 @@ export const CustomFormEdit = ({ closeCustomFormEdit }: CustomFormProps) => {
                                     textAreaRef.current?.value ?? ''
                                 setCustomFormSaving(true)
                                 setCustomFormError('')
+
+                                const data = formInput?.value?.id
+                                    ? {
+                                          htmlCode,
+                                          id: formId,
+                                      }
+                                    : {
+                                          htmlCode,
+                                          id: formId,
+                                          name: nameInput?.value,
+                                      }
                                 updateCustomForm(
-                                    operationType === 'create'
-                                        ? {
-                                              htmlCode,
-                                              id: formId,
-                                              name: nameInput?.value,
-                                          }
-                                        : {
-                                              htmlCode,
-                                              id: formId,
-                                          }
+                                    data,
+                                    onSuccess,
+                                    onError,
+                                    formInput?.value?.id
                                 )
                             }}
-                            dataTest="form-submit-button"
-                            disabled={customFormSaving}
                         >
                             {i18n.t('Save custom form')}
                         </Button>