+++ src/lib/form/index.ts
@@ -1,4 +1,5 @@
 export { usePatchModel } from './usePatchModel'
+export { useClearFormFields } from './useClearFormFields'
 export { composeAsyncValidators } from './composeAsyncValidators'
 export type { FormFieldValidator } from './composeAsyncValidators'
 export { required } from './validators'