+++ src/lib/sectionList/listViews/sectionListViewsConfig.ts
@@ -494,6 +494,22 @@ export const modelListViewsConfig = {
             overrideDefaultAvailable: true,
         },
     },
+    programRule: {
+        columns: {
+            default: [
+                DESCRIPTORS.name,
+                { label: i18n.t('Program name'), path: 'program.displayName' },
+                DESCRIPTORS.publicAccess,
+                'lastUpdated',
+            ],
+            available: ['description', 'priority', 'condition', 'programStage'],
+        },
+        filters: {
+            default: ['program'],
+            available: [],
+            overrideDefaultAvailable: true,
+        },
+    },
     optionGroupSet: {
         columns: {
             default: [
@@ -518,6 +534,14 @@ export const modelListViewsConfig = {
                 'fromToName',
                 'toFromName',
                 'referral',
+                {
+                    label: i18n.t('Initiating entity type'),
+                    path: 'fromConstraint.relationshipEntity',
+                },
+                {
+                    label: i18n.t('Receiving entity type'),
+                    path: 'toConstraint.relationshipEntity',
+                },
             ],
         },
         filters: {
@@ -725,6 +749,7 @@ export const modelListViewsConfig = {
                 { label: i18n.t('Option set'), path: 'optionSet.displayName' },
             ],
             available: [
+                { label: i18n.t('Sort order'), path: 'sortOrder' },
                 {
                     label: i18n.t('Category attribute'),
                     path: 'categoryAttribute',