+++ src/pages/programRules/form/index.ts
@@ -1,0 +1,1 @@
+export { programRuleListSchema } from './programRuleSchema'