+++ src/pages/programsWip/form/programStage/stageSchema.ts
@@ -1,8 +1,15 @@
 import { z } from 'zod'
 import { getDefaults, modelFormSchemas } from '../../../../lib'
+import { ProgramStage } from '../../../../types/generated'
 
-const { identifiable, modelReference, withAttributeValues, style } =
-    modelFormSchemas
+const {
+    identifiable,
+    modelReference,
+    withAttributeValues,
+    style,
+    withDefaultListColumns,
+    referenceCollection,
+} = modelFormSchemas
 
 export const stageSchema = identifiable.merge(withAttributeValues).extend({
     description: z.string().optional(),
@@ -18,6 +25,21 @@ export const stageSchema = identifiable.merge(withAttributeValues).extend({
     programStageLabel: z.string().optional(),
     eventLabel: z.string().optional(),
     program: modelReference,
+    programStageDataElements: referenceCollection.default([]),
+    repeatable: z.boolean().optional(),
+    standardInterval: z.number().optional(),
+    generatedByEnrollmentDate: z.boolean().optional(),
+    autoGenerateEvent: z.boolean().optional(),
+    openAfterEnrollment: z.boolean().optional(),
+    reportDateToUse: z.enum(['enrollmentDate', 'incidentDate']).optional(),
+    minDaysFromStart: z.number(),
+    hideDueDate: z.boolean().optional(),
+    periodType: z.nativeEnum(ProgramStage.periodType).optional(),
+    nextScheduleDate: modelReference.optional(),
+    allowGenerateNextVisit: z.boolean().optional(),
+    remindCompleted: z.boolean().optional(),
 })
+
+export const stageListSchema = stageSchema.merge(withDefaultListColumns)
 
 export const initialStageValue = getDefaults(stageSchema)