+++ src/pages/validationRules/form/ValidationRuleFormFields.tsx
@@ -22,7 +22,7 @@ import {
     StandardFormSectionTitle,
     MissingValueStrategyField,
 } from '../../../components'
-import { PaddedContainer } from '../../../components/metadataFormControls/ExpressionBuilder/PaddedContainer'
+import { PaddedContainer } from '../../../components/ExpressionBuilder/PaddedContainer'
 import {
     getConstantTranslation,
     SchemaName,
@@ -119,10 +119,7 @@ const ValidationRuleFormFields = () => {
                 </StandardFormSectionDescription>
 
                 <div className={css.expressionContainer}>
-                    <PaddedContainer>
-                        <div className={css.subtitle}>
-                            {i18n.t('Left side expression')}
-                        </div>
+                    <PaddedContainer title={i18n.t('Left side expression')}>
                         <StandardFormField>
                             <ExpressionBuilderEntry
                                 fieldName="leftSide.expression"
@@ -170,10 +167,7 @@ const ValidationRuleFormFields = () => {
                 </StandardFormField>
 
                 <div className={css.expressionContainer}>
-                    <PaddedContainer>
-                        <div className={css.subtitle}>
-                            {i18n.t('Right side expression')}
-                        </div>
+                    <PaddedContainer title={i18n.t('Right side expression')}>
                         <StandardFormField>
                             <ExpressionBuilderEntry
                                 fieldName="rightSide.expression"