+++ src/testUtils/builders.ts
@@ -58,8 +58,9 @@ import { predictorGroupListSchema } from '../pages/predictorGroups/form/predicto
 import { predictorListSchema } from '../pages/predictors/form/predictorSchema'
 import { programIndicatorGroupListSchema } from '../pages/programIndicatorGroups/form'
 import { programIndicatorsListSchema } from '../pages/programIndicators/form/programIndicatorsFormSchema'
+import { programRuleListSchema } from '../pages/programRules/form/programRuleSchema'
 import { programRuleVariableListSchema } from '../pages/programRuleVariables/form/programRuleVariableSchema'
-import { programListSchema } from '../pages/programsWip/form'
+import { programListSchema, stageListSchema } from '../pages/programsWip/form'
 import { relationshipTypeListSchema } from '../pages/relationshipTypes/form/RelationshipTypeFormSchema'
 import { trackedEntityAttributeListSchema } from '../pages/trackedEntityAttributes/form/TrackedEntityAttributeFormSchema'
 import { trackedEntityTypeListSchema } from '../pages/trackedEntityTypes/form'
@@ -352,6 +353,11 @@ export const testProgramRuleVariable = (overwrites: Record<any, any> = {}) => ({
     ...overwrites,
 })
 
+export const testProgramRule = (overwrites: Record<any, any> = {}) => ({
+    ...generateMock(programRuleListSchema, { mockeryMapper }),
+    ...overwrites,
+})
+
 export const testAttributeList = (overwrites: Record<any, any> = {}) => ({
     ...generateMock(attributeListSchema, { mockeryMapper }),
     ...overwrites,
@@ -482,6 +488,13 @@ export const testLegendSet = ({
 } = {}) => ({
     id,
     displayName,
+})
+
+export const testProgramStage = (overwrites: Record<any, any> = {}) => ({
+    ...generateMock(stageListSchema, {
+        mockeryMapper,
+    }),
+    ...overwrites,
 })
 
 export const testOrgUnitLevel = ({