+++ CHANGELOG.md
@@ -1,3 +1,109 @@
+## [0.122.2](https://github.com/dhis2/maintenance-app-beta/compare/v0.122.1...v0.122.2) (2026-02-04)
+
+
+### Bug Fixes
+
+* data element aggregation level in ou hierarchy ([6cbcce1](https://github.com/dhis2/maintenance-app-beta/commit/6cbcce15287599394618e7f95882d9cb262738b2))
+
+## [0.122.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.122.0...v0.122.1) (2026-02-03)
+
+
+### Bug Fixes
+
+* no quotes in css ([#756](https://github.com/dhis2/maintenance-app-beta/issues/756)) ([8e77f83](https://github.com/dhis2/maintenance-app-beta/commit/8e77f8397fefd29a583d60fa5008c9ff3969e824))
+
+# [0.122.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.121.0...v0.122.0) (2026-02-02)
+
+
+### Features
+
+* [DHIS2-20539] Program stage creation and scheduling ([#732](https://github.com/dhis2/maintenance-app-beta/issues/732)) ([8aea8eb](https://github.com/dhis2/maintenance-app-beta/commit/8aea8ebcc33ecee02dd2c6e26ce1fe5a799e62d8))
+
+# [0.121.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.120.1...v0.121.0) (2026-02-02)
+
+
+### Features
+
+* disable when configuring non supported TEAs as searchable ([1ad16d2](https://github.com/dhis2/maintenance-app-beta/commit/1ad16d217dbdcd63b4528fa953200e8c5848f29a))
+
+## [0.120.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.120.0...v0.120.1) (2026-02-02)
+
+
+### Bug Fixes
+
+* dataElements typo ([#754](https://github.com/dhis2/maintenance-app-beta/issues/754)) ([958f28c](https://github.com/dhis2/maintenance-app-beta/commit/958f28c1571ac6db81da0f4af71de06e821b203f))
+* impose sort order for attributes [DHIS2-13398] ([#737](https://github.com/dhis2/maintenance-app-beta/issues/737)) ([ad52174](https://github.com/dhis2/maintenance-app-beta/commit/ad52174a44f573cad9cd27cd0d2ce8b5d5bc66c4))
+
+# [0.120.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.119.0...v0.120.0) (2026-01-29)
+
+
+### Features
+
+* show initiating and receiving entity type in relationship type list ([ec317fc](https://github.com/dhis2/maintenance-app-beta/commit/ec317fc7bfba074781dba0fd1d94e74db4685790))
+
+# [0.119.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.118.0...v0.119.0) (2026-01-28)
+
+
+### Features
+
+* [DHIS2-20617] program rule variable form ([#729](https://github.com/dhis2/maintenance-app-beta/issues/729)) ([c11566a](https://github.com/dhis2/maintenance-app-beta/commit/c11566ac9e366f2e4efe05ae034a03afeabebf96))
+
+# [0.118.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.117.1...v0.118.0) (2026-01-28)
+
+
+### Features
+
+* reuse expression builder [DHIS2-20737] ([#735](https://github.com/dhis2/maintenance-app-beta/issues/735)) ([31fa406](https://github.com/dhis2/maintenance-app-beta/commit/31fa406b2fd44d5b12502c6776299e1fcfd470e0))
+
+## [0.117.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.117.0...v0.117.1) (2026-01-28)
+
+
+### Bug Fixes
+
+* program add new stage bug ([69f1294](https://github.com/dhis2/maintenance-app-beta/commit/69f1294c456fb42ede2cee328893f813068edea3))
+
+# [0.117.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.116.0...v0.117.0) (2026-01-28)
+
+
+### Features
+
+* program stage forms ([d3c2ecf](https://github.com/dhis2/maintenance-app-beta/commit/d3c2ecfc4eab82ca59a0eef24cc7349c496956b4))
+
+# [0.116.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.115.0...v0.116.0) (2026-01-28)
+
+
+### Features
+
+* [DHIS2-20537] Program stage data ([#733](https://github.com/dhis2/maintenance-app-beta/issues/733)) ([ff5212b](https://github.com/dhis2/maintenance-app-beta/commit/ff5212b6ffdde64b867b115f4826cd435c7b0a77))
+
+# [0.115.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.114.0...v0.115.0) (2026-01-28)
+
+
+### Features
+
+* adjust program stage list ([8141f52](https://github.com/dhis2/maintenance-app-beta/commit/8141f52206c5113921d7b970bb61705633ea266d))
+
+# [0.114.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.113.2...v0.114.0) (2026-01-27)
+
+
+### Features
+
+* add program rules list ([#738](https://github.com/dhis2/maintenance-app-beta/issues/738)) ([cb3e8ca](https://github.com/dhis2/maintenance-app-beta/commit/cb3e8caecf1fc1f958edcfa29f19a4bb85a3451c))
+
+## [0.113.2](https://github.com/dhis2/maintenance-app-beta/compare/v0.113.1...v0.113.2) (2026-01-27)
+
+
+### Bug Fixes
+
+* remove locales download ([#744](https://github.com/dhis2/maintenance-app-beta/issues/744)) ([017e26e](https://github.com/dhis2/maintenance-app-beta/commit/017e26e00c069d5b31cec93313f138369baaa373))
+
+## [0.113.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.113.0...v0.113.1) (2026-01-27)
+
+
+### Bug Fixes
+
+* keep whitespaces in expression builder validator ([#742](https://github.com/dhis2/maintenance-app-beta/issues/742)) ([063f1e6](https://github.com/dhis2/maintenance-app-beta/commit/063f1e6e0b14c8e3f5a5f7b197d1a7ede3b83a55))
+
 # [0.113.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.112.0...v0.113.0) (2026-01-26)